requirements.txt
absl-py
asttokens
attrs
audeer
audiofile
audmath
Brotli
certifi
cffi
charset-normalizer
comet-ml
configobj
contourpy
cycler
decorator
dulwich
everett
executing
filelock
fonttools
fsspec
graphviz
grpcio
h5py
hdf5storage
idna
imageio
imageio-ffmpeg
importlib_metadata
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
lazy_loader
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mkl_fft
mkl_random
mkl-service
moviepy
mpmath
natsort
networkx
numpy
opencv-python
packaging
pandas
parso
pexpect
pillow
pip
proglog
prompt_toolkit
protobuf
psutil
ptyprocess
pure_eval
py-cpuinfo
pycparser
pydub
Pygments
pymediainfo
pyparsing
PySocks
python-box
python-dateutil
python-dotenv
pytz
PyYAML
referencing
requests
requests-toolbelt
rich
rpds-py
scikit-image
scikit-learn
scipy
seaborn
semantic-version
sentry-sdk
setuptools
simplejson
six
soundfile
stack-data
sympy
tensorboard
tensorboard_data_server
tensorboardX
threadpoolctl
tifffile
torch
torch-tb-profiler
torchaudio
torchsummary
torchvision
torchviz
tqdm
traitlets
triton
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
wcwidth
Werkzeug
wheel
wrapt
wurlitzer
yacs
zipp
